+++ deser/tests/test_custom_map.rs
@@ -19,17 +19,17 @@ pub struct FlagsMapEmitter<'a> {
 }
 
 impl<'a> MapEmitter for FlagsMapEmitter<'a> {
-    fn next_key(&mut self, _state: &SerializerState) -> Option<SerializeHandle> {
-        if let Some((key, value)) = self.iter.next() {
+    fn next_key(&mut self, _state: &SerializerState) -> Result<Option<SerializeHandle>, Error> {
+        Ok(if let Some((key, value)) = self.iter.next() {
             self.value = Some(value);
             Some(SerializeHandle::boxed(key.to_string()))
         } else {
             None
-        }
+        })
     }
 
-    fn next_value(&mut self, _state: &SerializerState) -> SerializeHandle {
-        SerializeHandle::to(self.value.unwrap())
+    fn next_value(&mut self, _state: &SerializerState) -> Result<SerializeHandle, Error> {
+        Ok(SerializeHandle::to(self.value.unwrap()))
     }
 }